Complete exercise 08_removeFromArray.js
Completed the exercise through learning to use
"rest parameters" for functions. They allow a
function to accept an indefinite number of
arguments, which are then stored in an array.

Similar looking, but different, is spread
syntax (...). It allows you to expand an iterable
data type. This expanded form might become a
comma-seperated list, or an indexed list, or maybe
something else too?