Complete exercise 13_palindromes
Used a regex approach, but the solution didn't.
The solution used a combination of string methods
and array methods. The array methods allowed the
splitting and joining of the string
character-by-character. Additionally, they
allowed the use of ".filter()", to remove
punctuational characters from the string.

The ".filter()" method used a callback function
involving the "str.includes()" method. A string
named "alpahanumerical" stored characters [a-z]
and [0-9]. When running
"alphanumerical.includes(element)", it acts
similarly to a regex matching method.

Instead of using pointers to check if the
string is a palindrome, they simply reversed
the "cleaned" string and checked if it was
equal to the non-reversed string.

Also changed the local email used for git.